Revert "Update WallPaper.cpp"
This reverts commit f7966c608914a15d7479729c217d59ce2169d199.

+++ Source/WallPaper/Private/WallPaper.cpp
@@ -113,9 +113,13 @@ void FWallPaperModule::ShutdownModule()
 	WallpaperCommands::Unregister();
 	StyleSettings->ResetStyleColor();
 	//Clear Cache
-
+	FString PluginsPath = FPaths::ProjectPluginsDir()/"Wallpaper";
+	if(!IFileManager::Get().DirectoryExists(*PluginsPath))
+    	{
+    		PluginsPath = FPaths::EnginePluginsDir()/"Wallpaper";
+    	}
     		
-	FString TargetFilePath = FPaths::ProjectSavedDir()/"Cache";
+    	FString TargetFilePath = PluginsPath/"Content/Cache";
     	if(IFileManager::Get().DirectoryExists(*TargetFilePath))
     	{
     		IFileManager::Get().DeleteDirectory(*(TargetFilePath),false,true);
@@ -625,7 +629,7 @@ void FWallPaperModule::ImportPicTheme()
 		PluginsPath = FPaths::EnginePluginsDir()/"Marketplace/Wallpaper";
 	}
 		
-	FString TargetFilePath = FPaths::ProjectSavedDir()/"Cache";
+	FString TargetFilePath = PluginsPath/"Content/Cache";
 	//clear file
 	if(IFileManager::Get().DirectoryExists(*TargetFilePath))
 	{
